feat(refactor): разнес все обработчики по смыслу и их регистрацию в предзапуск бота
