+++ merc_gve/handler/__init__.py
@@ -1,0 +1,2 @@
+from .command import register_handlers_command
+from .mercury_subscribe import register_handlers_chcked_mercury
